requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Pillow==9.5.0
pytz==2023.3
Flask